src/main/python/requirements-to-freeze.txt
#torch
torchvision
seaborn
numpy
tabulate
sklearn
tensorboard
tqdm
pytest
